Drop items on break block

+++ src/main/java/be/thatsatim/gravestone/listeners/BreakBlock.java
@@ -2,16 +2,21 @@ package be.thatsatim.gravestone.listeners;
 
 import be.thatsatim.gravestone.Gravestone;
 import be.thatsatim.gravestone.database.GravestoneDatabase;
+import be.thatsatim.gravestone.utils.ItemSerializer;
 import be.thatsatim.gravestone.utils.Memory;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.ItemStack;
 
 import java.sql.SQLException;
+import java.util.UUID;
 
 public class BreakBlock implements Listener {
 
@@ -20,11 +25,39 @@ public class BreakBlock implements Listener {
     }
 
     @EventHandler
-    public void onBreakBlock(BlockBreakEvent event) {
+    public void onBreakBlock(BlockBreakEvent event) throws SQLException {
         Block block = event.getBlock();
         Location location = block.getLocation();
+        Player player = event.getPlayer();
 
         if (block.getType().equals(Material.MOSSY_STONE_BRICK_STAIRS)) {
+            if (!Memory.getGravestone(location)) { return; }
+
+            String[] values = GravestoneDatabase.getGravestone(location);
+
+            // UUID and player related logic
+            String graveOwnerString = values[0];
+            Player graveOwnerPlayer = Bukkit.getPlayer(UUID.fromString(graveOwnerString));
+
+            // Permission logic
+            if (player != graveOwnerPlayer && !player.hasPermission("MapleGrave.Staff")) {
+                event.setCancelled(true);
+                return;
+            }
+
+            // Inventory logic
+            String DatabaseInventory = values[1];
+            ItemStack[] inventory = ItemSerializer.deserializeInventory(DatabaseInventory, 0);
+
+            // Drop the inventory
+            if (inventory != null) {
+                for (ItemStack item : inventory) {
+                    if (item != null) {
+                        block.getWorld().dropItemNaturally(location, item);
+                    }
+                }
+            }
+
             try {
                 GravestoneDatabase.deleteGravestone(location);
                 Memory.deleteGravestone(location);
